feat: split forest cli and daemon (#1973)
