+++ Cargo.toml
@@ -1,6 +1,8 @@
 [workspace]
 members = [
-  "forest",
+  "forest/cli",
+  "forest/daemon",
+  "forest/shared",
   "blockchain/beacon",
   "blockchain/blocks",
   "blockchain/chain",
@@ -116,6 +118,7 @@ forest_beacon          = { path = "./blockchain/beacon" }
 forest_blocks          = { path = "./blockchain/blocks" }
 forest_chain           = { path = "./blockchain/chain" }
 forest_chain_sync      = { path = "./blockchain/chain_sync" }
+forest_cli_shared      = { path = "./forest/shared" }
 forest_crypto          = { path = "./crypto" }
 forest_db              = { path = "./node/db" }
 forest_deleg_cns       = { path = "./blockchain/consensus/deleg_cns" }
