+++ forest/cli/src/cli/auth_cmd.rs
