+++ forest/daemon/src/cli/mod.rs
@@ -1,0 +1,48 @@
+// Copyright 2019-2022 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use forest_cli_shared::cli::{CliOpts, FOREST_VERSION_STRING};
+use futures::channel::oneshot::Receiver;
+use log::{info, warn};
+use std::cell::RefCell;
+use std::process;
+use std::sync::atomic::{AtomicUsize, Ordering};
+use std::sync::Arc;
+use structopt::StructOpt;
+
+/// CLI structure generated when interacting with Forest binary
+#[derive(StructOpt)]
+#[structopt(
+    name = env!("CARGO_PKG_NAME"),
+    version = FOREST_VERSION_STRING.as_str(),
+    about = env!("CARGO_PKG_DESCRIPTION"),
+    author = env!("CARGO_PKG_AUTHORS")
+)]
+pub struct Cli {
+    #[structopt(flatten)]
+    pub opts: CliOpts,
+    pub cmd: Option<String>,
+}
+
+pub fn set_sigint_handler() -> Receiver<()> {
+    let (ctrlc_send, ctrlc_oneshot) = futures::channel::oneshot::channel();
+    let ctrlc_send_c = RefCell::new(Some(ctrlc_send));
+
+    let running = Arc::new(AtomicUsize::new(0));
+    ctrlc::set_handler(move || {
+        let prev = running.fetch_add(1, Ordering::SeqCst);
+        if prev == 0 {
+            warn!("Got interrupt, shutting down...");
+            // Send sig int in channel to blocking task
+            if let Some(ctrlc_send) = ctrlc_send_c.try_borrow_mut().unwrap().take() {
+                ctrlc_send.send(()).expect("Error sending ctrl-c message");
+            }
+        } else {
+            info!("Exiting process");
+            process::exit(0);
+        }
+    })
+    .expect("Error setting Ctrl-C handler");
+
+    ctrlc_oneshot
+}
