+++ forest/daemon/src/daemon.rs
@@ -1,13 +1,13 @@
 // Copyright 2019-2022 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-use super::cli::{set_sigint_handler, Config, FOREST_VERSION_STRING};
-use crate::cli_error_and_die;
+use super::cli::set_sigint_handler;
 use async_std::{channel::bounded, net::TcpListener, task};
 use forest_auth::{create_token, generate_priv_key, ADMIN, JWT_IDENTIFIER};
 use forest_chain::ChainStore;
 use forest_chain_sync::consensus::SyncGossipSubmitter;
 use forest_chain_sync::ChainMuxer;
+use forest_cli_shared::cli::{cli_error_and_die, Config, FOREST_VERSION_STRING};
 use forest_db::rocks::RocksDb;
 use forest_fil_types::verifier::FullVerifier;
 use forest_genesis::{get_network_name_from_genesis, import_chain, read_genesis_header};
