+++ forest/shared/Cargo.toml
@@ -1,0 +1,76 @@
+[package]
+name        = "forest_cli_shared"
+version     = "0.4.1"
+authors     = ["ChainSafe Systems <info@chainsafe.io>"]
+description = "Filecoin implementation in Rust. This crate contains shared utilities that power both cli and daemon."
+edition     = "2021"
+
+[dependencies]
+anyhow.workspace                 = true
+async-log                        = "2.0"
+async-std                        = { workspace = true, features = ["attributes", "tokio1"] }
+base64.workspace                 = true
+cid                              = { workspace = true, default-features = false, features = ["std"] }
+directories                      = "4.0.1"
+fil_actor_miner_v8.workspace     = true
+forest_actor_interface.workspace = true
+forest_beacon.workspace          = true
+forest_blocks.workspace          = true
+forest_chain.workspace           = true
+forest_chain_sync.workspace      = true
+forest_crypto.workspace          = true
+forest_db.workspace              = true
+forest_deleg_cns                 = { workspace = true, optional = true }
+forest_encoding.workspace        = true
+forest_fil_types.workspace       = true
+forest_interpreter.workspace     = true
+forest_ipld_blockstore.workspace = true
+forest_json.workspace            = true
+forest_key_management.workspace  = true
+forest_libp2p.workspace          = true
+forest_message                   = { workspace = true, features = ["blst"] }
+forest_message_pool.workspace    = true
+forest_networks.workspace        = true
+forest_rpc-api.workspace         = true
+forest_rpc-client.workspace      = true
+forest_state_manager.workspace   = true
+forest_utils.workspace           = true
+futures.workspace                = true
+fvm.workspace                    = true
+fvm_ipld_car.workspace           = true
+fvm_ipld_encoding.workspace      = true
+fvm_shared                       = { workspace = true, default-features = false }
+git-version                      = "0.3"
+hex.workspace                    = true
+lazy_static.workspace            = true
+libp2p                           = { workspace = true, default-features = false, features = ["identify"] }
+log.workspace                    = true
+multibase.workspace              = true
+num-rational.workspace           = true
+num-traits.workspace             = true
+once_cell.workspace              = true
+pbr.workspace                    = true
+pretty_env_logger.workspace      = true
+prometheus                       = { workspace = true, features = ["process"] }
+rayon.workspace                  = true
+reqwest                          = { version = "0.11", features = ["stream"] }
+serde                            = { workspace = true, features = ["derive"] }
+serde_json.workspace             = true
+sha2                             = "0.10.5"
+structopt                        = "0.3"
+tempfile                         = "3.3"
+time                             = "0.3"
+tokio                            = { workspace = true, features = ["sync"] }
+toml                             = "0.5"
+
+
+[dependencies.jsonrpc-v2]
+default-features = false
+features         = ["easy-errors", "macros", "bytes-v05"]
+version          = "0.11"
+
+[dev-dependencies]
+assert_cmd        = "2"
+quickcheck        = "1"
+quickcheck_macros = "1"
+rand.workspace    = true
